Update loop variable types from int to size_t
This change ensures compatibility in situations where the array size may exceed the limits of an int. Additionally, it standardizes the loop variable type across various unit tests for vectors, linked lists, stacks, and doubly linked lists for consistency and correctness. Also, added AddressSanitizer flags to CMakeLists for enhanced debugging.

+++ C/test/assert_doubly_linked_list_lib.c
@@ -8,7 +8,7 @@ void test_int_doubly_linked_list() {
     const int data[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
     dll_node *root = dll_new_node(&data[0]);
     dll_node *curr = root;
-    for (int i = 1; i < sizeof(data) / sizeof(data[0]); i++) {
+    for (size_t i = 1; i < sizeof(data) / sizeof(data[0]); i++) {
         curr = dll_insert_after_node(curr, &data[i]);
     }
 
@@ -31,7 +31,7 @@ void test_int_dll_random_inserts() {
     dll_insert_before_node(root->prev, &data[0]);
 
     dll_node const *curr = root->prev->prev;
-    for (int i = 0; i < sizeof(data) / sizeof(data[0]); i++) {
+    for (size_t i = 0; i < sizeof(data) / sizeof(data[0]); i++) {
         assert(*(int *) curr->data == data[i]);
         curr = curr->next;
     }
@@ -48,7 +48,7 @@ void test_string_dll_random_inserts() {
     dll_insert_before_node(root->prev, &data[0]);
 
     dll_node const *curr = root->prev->prev;
-    for (int i = 0; i < sizeof(data) / sizeof(data[0]); i++) {
+    for (size_t i = 0; i < sizeof(data) / sizeof(data[0]); i++) {
         assert(*(char **) curr->data == data[i]);
         curr = curr->next;
     }
